rename site settings file
