+++ application/dataentry/views/__init__.py
@@ -1,4 +1,4 @@
 from all_views import *
 from photo_exporter import *
 from addresses import *
-from SiteSettings import *
+from site_settings import *
